provided sensible names for SuitesCreator subclasses to avoid clash with SuitesProvider subclasses
