+++ compiler/src/org.graalvm.compiler.core.aarch64/src/org/graalvm/compiler/core/aarch64/AArch64SuitesCreator.java
@@ -22,13 +22,14 @@
  */
 package org.graalvm.compiler.core.aarch64;
 
-import org.graalvm.compiler.java.DefaultSuitesProvider;
+import org.graalvm.compiler.java.DefaultSuitesCreator;
 import org.graalvm.compiler.nodes.graphbuilderconf.GraphBuilderConfiguration.Plugins;
 import org.graalvm.compiler.phases.tiers.CompilerConfiguration;
 
-public class AArch64SuitesProvider extends DefaultSuitesProvider {
+public class AArch64SuitesCreator extends DefaultSuitesCreator
+{
 
-    public AArch64SuitesProvider(CompilerConfiguration compilerConfiguration, Plugins plugins) {
+    public AArch64SuitesCreator(CompilerConfiguration compilerConfiguration, Plugins plugins) {
         super(compilerConfiguration, plugins);
     }
